+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.0.2] - 2018-02-17
+
 ### Fixed
 - Now library uses function types instead of interface type
 
@@ -33,5 +35,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add documentations for functions.
 
 [Unreleased]: https://github.com/Ailrun/tsdux
+[1.0.2]: https://github.com/Ailrun/tsdux/tree/v1.0.2
 [1.0.1]: https://github.com/Ailrun/tsdux/tree/v1.0.1
 [1.0.0]: https://github.com/Ailrun/tsdux/tree/v1.0.0